A new commit after merging some local stashed changes again.

+++ src/main/java/kaica_dun/entities/Dungeon.java
@@ -6,7 +6,7 @@ import java.util.*;
 @Entity
 @Table(name = "dungeon")
 public class Dungeon {
-    private int dungeonId;
+    private Long dungeonId;
     private Player player;
     private int roomRows;
     private int roomColumns;
@@ -32,11 +32,11 @@ public class Dungeon {
 
     @Id @GeneratedValue
     @Column(name = "dungeonID")
-    public int getDungeonId() {
+    public Long getDungeonId() {
         return dungeonId;
     }
 
-    public void setDungeonId(int dungeonId) { this.dungeonId = dungeonId; }
+    public void setDungeonId(Long dungeonId) { this.dungeonId = dungeonId; }
 
     @OneToMany(cascade = { CascadeType.PERSIST,
                            CascadeType.MERGE,
