+++ src/main/java/kaica_dun/entities/Item.java
@@ -6,7 +6,7 @@ import java.util.Objects;
 @Entity
 @Table(name = "item")
 public class Item {
-    private int itemId;
+    private Long itemId;
     private String itemName;
     private String description;
     private Integer playerId;
@@ -15,11 +15,11 @@ public class Item {
 
     @Id
     @Column(name = "itemID")
-    public int getItemId() {
+    public Long getItemId() {
         return itemId;
     }
 
-    public void setItemId(int itemId) {
+    public void setItemId(Long itemId) {
         this.itemId = itemId;
     }
